Add simple optimizations
If we have 3 souls and none are the primary set, then we won't be able
to construct a primary set so skip this start. Testing at 4 and 5 are
harder because we have several combinations we have to check.

Also move the test for at least 4 of the primary earlier, and put it in
EachSet to group the primary tests in the same function.